푸시 알림 클릭 후 화면 라우팅 이후 홈화면 자동전환 : refactor : app_navigator 삭제, import 정리 https://github.com/TEAM-ROMROM/RomRom-FE/issues/827
